fix: rename hook and improve helpers
