+++ store/songsStore.ts
@@ -88,14 +88,14 @@ export const useMusicStore = create<MusicStoreState>((set, get) => ({
 		});
 	},
 
-	getSongById: async (id) => {
-		try {
-			return await database.get<Song>("songs").find(id);
-		} catch (e) {
-			console.error("getSongById error", e);
-			return null;
-		}
-	},
+        getSongById: async (id) => {
+                try {
+                        return await database.get<Song>("songs").find(id);
+                } catch (e) {
+                        console.error("getSongById error", e);
+                        return null;
+                }
+        },
 
 	updateSong: async (id, updates) => {
 		await database.write(async () => {
@@ -184,15 +184,15 @@ export const useMusicStore = create<MusicStoreState>((set, get) => ({
 		set({ artists, isLoading: false });
 	},
 
-	getArtistById: async (id) => {
-		try {
-			const artist = await database.get<Artist>("artists").find(id);
-			return artist;
-		} catch (e) {
-			console.error("getSongById error", e);
-			return null;
-		}
-	},
+        getArtistById: async (id) => {
+                try {
+                        const artist = await database.get<Artist>("artists").find(id);
+                        return artist;
+                } catch (e) {
+                        console.error("getArtistById error", e);
+                        return null;
+                }
+        },
 
 	setPlayingSongId: (id) => {
 		set({ playingSongId: id });
