+++ utils/formatSeconds.ts
@@ -1,5 +1,5 @@
 export const formatSeconds = (seconds: number): string => {
   const minutes = Math.floor(seconds / 60);
-  const remainingSeconds = seconds % 60;
-  return `${minutes}:${Math.floor(remainingSeconds).toString().padStart(2, "0")}`;
-}
+  const remainingSeconds = Math.floor(seconds % 60);
+  return `${minutes}:${remainingSeconds.toString().padStart(2, "0")}`;
+};
